+++ app/controllers/dashboards_controller.rb
@@ -8,7 +8,7 @@ class DashboardsController < ApplicationController
   # Displays the index view with the default dashboard
   def index
     @accessible_projects = current_user.accessible_projects
-    @most_active_projects = Project.order("updated_at DESC").limit(5)
+    @most_active_projects = Project.most_active(5)
     @audits = Audited::Adapters::ActiveRecord::Audit.unscoped.where(:associated_id => [@accessible_projects.map { |x| x.id }], :associated_type => Project.name).order("created_at DESC").limit(25)
   end
 end