+++ app/controllers/sprints_controller.rb
@@ -3,9 +3,9 @@ class SprintsController < AdminController
   self.responder = ::ApplicationResponder
   respond_to :html, :json, :pjax
 
-  before_filter :require_login, :except => [:current]
   before_filter :load_user
   before_filter :load_project
+  before_filter :require_login, :except => [:current]
   before_filter :ensure_can_collaborate, :except => [:current]
 
   helper :issues