+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # Changelog
 
-## 0.6.2
+## v0.7.1
+
+* Polishes report list view
+
+## v0.7.0
+
+* Adds support for private projects
+
+## v0.6.2
 
 * Growable Textareas enabled (projects, issues)